Se pintan aquellos articulos que se pidieron

+++ js/articulo.js
@@ -1,11 +1,13 @@
 function crearArticulo(nombre, precio_bulto, precio_unidad, empaque, cantidad){
     var precio;
     var parcial;
+
     if (empaque === 'Bulto'){
-      precio = parseInt(precio_bulto);
+
+      precio = parseFloat(precio_bulto);
     }
     else {
-      precio = parseInt(precio_unidad);
+      precio = parseFloat(precio_unidad);
     }
     parcial = parseInt(cantidad) * precio;
